turned 'get_policy' into a function to remove duplicate code

+++ src/lavatory/commands/purge.py
@@ -6,7 +6,7 @@ import click
 
 from ..utils.artifactory import Artifactory
 from ..utils.performance import get_performance_report
-from ..utils.setup_pluginbase import setup_pluginbase
+from ..utils.setup_pluginbase import setup_pluginbase, get_policy
 
 LOG = logging.getLogger(__name__)
 
@@ -52,17 +52,10 @@ def apply_purge_policies(selected_repos, policies_path=None, dryrun=True, defaul
     plugin_source = setup_pluginbase(extra_policies_path=policies_path)
     LOG.info("Applying retention policies to %s", ', '.join(selected_repos))
     for repository in selected_repos:
-        policy_name = repository.replace("-", "_")
         artifactory_repo = Artifactory(repo_name=repository)
-        try:
-            policy = plugin_source.load_plugin(policy_name)
-        except ImportError:
-            if default:
-                LOG.info("No policy found for %s. Applying Default", repository)
-                policy = plugin_source.load_plugin('default')
-            else:
-                LOG.info("No policy found for %s. Skipping Default", repository)
-                continue
+        policy = get_policy(plugin_source, repository, default=default)
+        if not policy:
+            continue
         LOG.info("Policy Docs: %s", inspect.getdoc(policy.purgelist))
         artifacts = policy.purgelist(artifactory_repo)
         purged_count = artifactory_repo.purge(dryrun, artifacts)
